+++ Retia.Tests/Mathematics/MatrixTests.Data.cs
@@ -1,4 +1,4 @@
-﻿using MathNet.Numerics.LinearAlgebra.Single;
+﻿using MathNet.Numerics.LinearAlgebra.Double;
 using Retia.Mathematics;
 
 namespace Retia.Tests.Mathematics
